Fix a self-inflicted count truncation, RedisClusterClient bugs, mislabeled cluster tests, and finish decoder relocation
Round 4 of the full-repo release-readiness review found: bzmPopMin/bzmPopMax
silently truncating count to Int (a regression from this branch's own
DurationOps.toSecondsOrZero fix shifting Lettuce overload resolution) - fixed by
changing count's type to Int, matching the already-Int zmPopMin/zmPopMax;
RedisClusterClient.nodeId/partitions crashing/aliasing Lettuce's live mutable
state; four RedisClusterSpec "cluster: ..." tests that were silently running
against a single-node instance, which once fixed surfaced two genuine
cluster-compatibility bugs (a CROSSSLOT-unsafe hyperloglog test key layout, and
FUNCTION LOAD's non-cluster-wide semantics). Also relocates the last six
BaseRedis-private decoders (GeoSearchResult, StreamMessage,
StreamEntryDeletionResult, XPendingSummary, XPendingMessage, XAutoClaimResult) to
their companion objects, and adds missing zScore/zRank/zRevRank test coverage.

+++ modules/effects/src/main/scala/dev/profunktor/redis4cats/algebra/sortedsets.scala
@@ -64,8 +64,8 @@ trait SortedSetGetter[F[_], K, V] {
 
   /** Pops up to `count` members with the highest scores from the first non-empty of `keys`. */
   def zmPopMax(keys: NonEmptyList[K], count: Int): F[Option[(K, List[ScoreWithValue[V]])]]
-  def bzmPopMin(timeout: Duration, keys: NonEmptyList[K], count: Long): F[Option[(K, List[ScoreWithValue[V]])]]
-  def bzmPopMax(timeout: Duration, keys: NonEmptyList[K], count: Long): F[Option[(K, List[ScoreWithValue[V]])]]
+  def bzmPopMin(timeout: Duration, keys: NonEmptyList[K], count: Int): F[Option[(K, List[ScoreWithValue[V]])]]
+  def bzmPopMax(timeout: Duration, keys: NonEmptyList[K], count: Int): F[Option[(K, List[ScoreWithValue[V]])]]
   def zUnion(args: Option[ZAggregateArgs], keys: K*): F[List[V]]
   def zUnionWithScores(args: Option[ZAggregateArgs], keys: K*): F[List[ScoreWithValue[V]]]
   def zInter(args: Option[ZAggregateArgs], keys: K*): F[List[V]]
